# Avoid allocations in global objects