It is potentially unsafe (especially when using dynamic libraries).