@lemire 2020-07-30
Merged.